requirements.txt
streamlit
torch
open_clip_torch
pillow
pandas
transformers
matplotlib
peft
joblib
scikit-learn
opencv-python-headless
